Do not trim after expanding ranges
Here we expand ranges with an arbitrary value, merge overlapping
ranges, and shrink by the same arbitrary value, to merge neighbouring
ranges distant of at most that value.  The trimming operation that
is being removed by this patch was breaking the algorithm: some isolated
ranges near sequence feature (chromosomes, contigs, ...) boundaries would
not have their original coordinates after the operation.

The broken function was tested on the Okinawa genes, and passed the
test by chance, but then the full rebuild failed when processing
the Osaka genes…